Remove dangerous endpoints
The questions are required to not be revealed outside a
quiz session. Creating a quiz questsion will be accompanid
with the creation of the quiz itself. No questions can be updated,
to prevent the incosistency between user results and experience.
Thus, keeping these endpoints make no sense. In fact, it's dangerous,
because if the users are allowed to know the questions beforehead,
cheating will become easy.