Remove dangerous endpoints
The answers are required to not be revealed outside a
quiz session. Creating a question answer will be accompanid
with the creation of the quiz itself. No answers can be neither
updated nor deleted, to prevent incosistencies between user scores
and experience. Thus, keeping these endpoints make no sense. In fact,
it's dangerous, because if the users are allowed to know the answers
beforehead, cheating will become easy.