Start a quiz taking session in redis-stack-server
This is one of the most important methods in the entire
application. This method checks if a certain quiz is
taken by another user or not, if yes, it just links the
current user's ongoing quiz session to that quiz to avoid
wasting resources, and if not, it creates a new ongoing quiz
session in the cache.

It also populates a set of session names to track the number
of sessions that are linked to the target quiz for removal
of quiz when no session is associated with it.